pyproject.toml
[project]
name = "meshbot"
version = "0.1.0"
description = "A MeshCore mesh network chatbot with AI agent support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "meshcore",
    "pydantic-ai",
    "fastmcp",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "anthropic",
    "openai",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
meshbot = "meshbot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=meshbot --cov-report=term-missing"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
